chore: 🐝 Update SDK - Generate 0.73.2

+++ src/models/operations/gettags.ts
@@ -52,7 +52,7 @@ export type GetTagsRequest = {
    */
   ids?: string | Array<string> | undefined;
   /**
-   * The page number for pagination.
+   * The page number for pagination. The first page is `1`.
    */
   page?: number | undefined;
   /**
@@ -102,8 +102,8 @@ export const GetTagsRequest$outboundSchema: z.ZodType<
   sortOrder: QueryParamSortOrder$outboundSchema.default("asc"),
   search: z.string().optional(),
   ids: z.union([z.string(), z.array(z.string())]).optional(),
-  page: z.number().optional(),
-  pageSize: z.number().default(100),
+  page: z.number().int().optional(),
+  pageSize: z.number().int().default(100),
 });
 
 export function getTagsRequestToJSON(getTagsRequest: GetTagsRequest): string {
