+++ src/models/operations/listfolders.ts
@@ -10,7 +10,7 @@ export type ListFoldersRequest = {
    */
   search?: string | undefined;
   /**
-   * The page number for pagination.
+   * The page number for pagination. The first page is `1`.
    */
   page?: number | undefined;
   /**
@@ -33,8 +33,8 @@ export const ListFoldersRequest$outboundSchema: z.ZodType<
   ListFoldersRequest
 > = z.object({
   search: z.string().optional(),
-  page: z.number().optional(),
-  pageSize: z.number().default(50),
+  page: z.number().int().optional(),
+  pageSize: z.number().int().default(50),
 });
 
 export function listFoldersRequestToJSON(
