+++ src/models/operations/listpartners.ts
@@ -97,11 +97,11 @@ export type ListPartnersRequest = {
    */
   tenantId?: string | undefined;
   /**
-   * A search query to filter partners by ID, name, email, or link.
+   * A search query to filter partners by ID, name, email, or company name.
    */
   search?: string | undefined;
   /**
-   * The page number for pagination.
+   * The page number for pagination. The first page is `1`.
    */
   page?: number | undefined;
   /**
@@ -603,8 +603,8 @@ export const ListPartnersRequest$outboundSchema: z.ZodType<
   email: z.string().optional(),
   tenantId: z.string().optional(),
   search: z.string().optional(),
-  page: z.number().optional(),
-  pageSize: z.number().default(100),
+  page: z.number().int().optional(),
+  pageSize: z.number().int().default(100),
 });
 
 export function listPartnersRequestToJSON(
